Move juggle components to juggleApp folder
