+++ src/components/juggleApp/AddTrick.jsx
