+++ src/index.js
@@ -1,16 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import reportWebVitals from "./reportWebVitals";
-import store from "./store/store";
-import App from "./App";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { Provider } from "react-redux";
+import store from "./store/store";
+import reportWebVitals from "./reportWebVitals";
+import App from "./App";
 import Home from "./components/pages/Home";
-import AddTrick from "./components/pages/AddTrick";
-import Chart from "./components/pages/Chart";
-import LearnToJuggle from "./components/pages/LearnToJuggle";
+import AddTrick from "./components/juggleApp/AddTrick";
+import Chart from "./components/juggleApp/Chart";
+import LearnToJuggle from "./components/juggleApp/LearnToJuggle";
+import TrickPage from "./components/juggleApp/TrickPage";
 import TaskApp from "./components/taskApp/TaskApp";
-import TrickPage from "./components/pages/TrickPage";
 import "./index.css";
 import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
